code some thing

+++ Backend/src/main/java/vn/edu/huce/beforeigner/infrastructures/exammodule/dtos/CompletedLessonDto.java
@@ -7,6 +7,6 @@ public class CompletedLessonDto {
 
     private Integer historyId;
 
-    private Float accuracy;
+    private Integer accuracy;
     
 }
